[sailfish-accounts-nextcloud] Handle failed user purges. JB#49870
Normally when an account is deleted, the user/albums/photos for that
that account are also deleted by the image syncer. However this mail
fail sometimes e.g. if msyncd has crashed. To guard against this,
purge any accounts as necessary when the image syncer is run. Also,
change the gallery UserModel so that it just loads users with
matching accounts, instead of failing to load when an account is not
present as expected.