[sailfish-accounts-nextcloud] Do image delta syncs whenever possible and update database progressively. JB#49870
Instead of doing a complete sync of the remote directory tree whenever
an image sync is triggered, only update the albums and photos that
have been modified according to the server etag for that resource,
which only changes on the server when the resource has changed.

Also, fetch and save albums one at a time so that each album can be
viewed in Gallery before the sync fully finishes. The WebDAV request
'Depth' is already set to '1' so it won't fetch more than one dir's
contents at a time. When each directory is fetched, compare it to the
local db contents and update the db immediately without waiting for
the sync to finish beforehand.