[account-nextcloud] Remove backup and sharing from sync schedule. JB#52335
Backup and sharing were already skipped when setting the sync schedule
for an already created account. However, when creating an account for
the first time, they were incorrectly having their sync schedule set to
match that of the other services.

This change ensures they are skipped (and therefore have sync schedule
disabled) when the account is first created.